Addressed feedback on XRInput from Jordan
